% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellAbort.R
\name{cromwell_abort}
\alias{cromwell_abort}
\title{Abort a workflow job on Cromwell}
\usage{
cromwell_abort(workflow_id)
}
\arguments{
\item{workflow_id}{(character) A single workflow ID}
}
\value{
Returns the response from the API post
}
\description{
Abort a workflow job on Cromwell
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwellSettings}
}

\author{
Amy Paguirigan, Scott Chamberlain
}
