% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellCall.R
\name{cromwell_call}
\alias{cromwell_call}
\title{Retrieve metadata for the calls made in a Cromwell workflow job}
\usage{
cromwell_call(workflow_id)
}
\arguments{
\item{workflow_id}{(character) A single workflow ID}
}
\value{
Returns a long form data frame of metadata on calls.
NOTE: does not currently support subWorkflows well yet.
}
\description{
Retrieve and process call metadata for a specific workflow.
}
\details{
Now supports nested scatters.
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwellSettings}
}

\examples{
\dontrun{
## Request what jobs have been submitted to your Cromwell instance in the
## past 7 days.
recentJobs <- cromwell_jobs(days = 7)
## Request workflow metadata for a specific job that was run in your
## Cromwell instance.
thisWorkflowID <- recentJobs$workflow_id[1]
callsMeta <- cromwell_call(workflow_id = thisWorkflowID)
}
}
\author{
Amy Paguirigan
}
