% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{cromwell_config}
\alias{cromwell_config}
\title{Cromwell Configuration}
\usage{
cromwell_config(cromwell_url = NULL, verbose = TRUE)
}
\arguments{
\item{cromwell_url}{(character) Cromwell server URL. The full url,
for example "http://localhost:8000" or "https://mycromwell.com"}

\item{verbose}{(logical) verbose messaging. default: \code{TRUE}. this
toggles on (\code{TRUE}) or off (\code{FALSE}) all messages throughout the package.}
}
\description{
Cromwell Configuration
}
\details{
Note that although the default value for \code{cromwell_url} is
\code{NULL}, we fetch the current value of \code{CROMWELLURL} env var internally
if \code{cromwell_url=NULL} - thus, we retain a value for \code{CROMWELLURL} if
it is set even if this function is run without passing anything to
\code{cromwell_url}
}
\examples{
\dontrun{
cromwell_config()
cromwell_config("https://mycromwellinstance.com")
cromwell_config(verbose = FALSE)
}
}
\author{
Scott Chamberlain
}
