% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellFailures.R
\name{cromwell_failures}
\alias{cromwell_failures}
\title{Pull metadata for the failed calls made in a Cromwell workflow job}
\usage{
cromwell_failures(workflow_id)
}
\arguments{
\item{workflow_id}{(character) A single workflow ID}
}
\value{
Returns a long form data frame of metadata on failed calls in a
workflow.
}
\description{
Gets info about failed calls for a specific workflow
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwellSettings}
}

\examples{
\dontrun{
## Request what jobs have been submitted to your Cromwell instance in the
## past 7 days.
recentJobs <- cromwell_jobs(days = 7)
## Request workflow metadata for a specific job that was run in your
## Cromwell instance.
thisWorkflowID <- recentJobs$workflow_id[1]
failsMeta <- cromwell_failures(workflow_id = thisWorkflowID)
}
}
\author{
Amy Paguirigan
}
