% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellGlob.R
\name{cromwell_glob}
\alias{cromwell_glob}
\title{Pull a glob of metadata for a specific Cromwell workflow job}
\usage{
cromwell_glob(
  workflow_id,
  expand_sub_workflows = FALSE,
  url = cw_url(),
  token = NULL
)
}
\arguments{
\item{workflow_id}{(character) A single workflow ID}

\item{expand_sub_workflows}{Boolean, whether to expand subworkflows in the
results or not, default is FALSE.}

\item{url}{(character) base url for your Cromwell server. optional. if not
supplied set the url as the env var \code{CROMWELLURL}}

\item{token}{(character) we do not recommend passing your token
here as a string. Either pass it using \code{\link[=Sys.getenv]{Sys.getenv()}} or save your
token as the env var \code{PROOF_TOKEN} and then passing nothing to this
param and we'll find it}
}
\value{
Returns a gross list of lists of metadata on a workflow.
}
\description{
Retrieve a glob of workflow level metadata for a specific workflow.
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwell_settings}
}

\examples{
\dontrun{
## Request what jobs have been submitted to your Cromwell instance in the
## past 7 days.
recentJobs <- cromwell_jobs(days = 7)
## Request workflow metadata for a specific job that was run in your
## Cromwell instance.
thisWorkflowID <- recentJobs$workflow_id[1]
workflowMeta <- cromwell_glob(workflow_id = thisWorkflowID)
}
}
\author{
Amy Paguirigan, Scott Chamberlain
}
