% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellJobs.R
\name{cromwell_jobs}
\alias{cromwell_jobs}
\title{Requests metadata about Cromwell workflow jobs during a time period specified}
\usage{
cromwell_jobs(
  days = 1,
  workflow_name = NULL,
  workflow_status = NULL,
  url = cw_url(),
  token = NULL
)
}
\arguments{
\item{days}{The number of days of history to return, defaults to 1 day.}

\item{workflow_name}{An array of strings of valid workflow names you want in
your job list.}

\item{workflow_status}{A array of strings of valid workflow statuses you want
in your job list (e.g., submitted, running, succeeded, failed,
aborting, aborted)}

\item{url}{(character) base url for your Cromwell server. optional. if not
supplied set the url as the env var \code{CROMWELLURL}}

\item{token}{(character) we do not recommend passing your token
here as a string. Either pass it using \code{\link[=Sys.getenv]{Sys.getenv()}} or save your
token as the env var \code{PROOF_TOKEN} and then passing nothing to this
param and we'll find it}
}
\value{
a tibble of metadata on workflow jobs submitted; each row is a
different job. columns:
\itemize{
\item end (dttm)
\item workflow_id (chr)
\item metadataArchiveStatus (chr)
\item workflow_name (chr)
\item start (dttm)
\item status (chr)
\item submission (dttm)
\item workflowDuration (dbl)
}
}
\description{
Requests metadata about Cromwell workflow jobs during a time period specified
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwell_settings}
}

\examples{
\dontrun{
## Request what jobs have been submitted to your Cromwell instance in the
## past 7 days.
recentJobs <- cromwell_jobs(days = 7)
}
}
\author{
Amy Paguirigan, Scott Chamberlain
}
