% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellValidate.R
\name{cromwell_validate}
\alias{cromwell_validate}
\title{Validates a workflow for submission}
\usage{
cromwell_validate(wdl, all_inputs = NULL, url = cw_url(), token = NULL)
}
\arguments{
\item{wdl}{Local path to the wdl file describing the workflow. (Required)}

\item{all_inputs}{Local path to the json containing ALL the inputs the
parameters to use with the workflow - if a batch and parameters json
are present they must first be combined. (Optional)}

\item{url}{(character) base url for your Cromwell server. optional. if not
supplied set the url as the env var \code{CROMWELLURL}}

\item{token}{(character) we do not recommend passing your token
here as a string. Either pass it using \code{\link[=Sys.getenv]{Sys.getenv()}} or save your
token as the env var \code{PROOF_TOKEN} and then passing nothing to this
param and we'll find it}
}
\value{
a list with validation details
}
\description{
Validates a workflow for submission
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwell_settings}
}

\author{
Amy Paguirigan, Scott Chamberlain
}
