% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellVersion.R
\name{cromwell_version}
\alias{cromwell_version}
\title{Get the version of a Cromwell server}
\usage{
cromwell_version(url = cw_url(), token = NULL)
}
\arguments{
\item{url}{(character) base url for your Cromwell server. optional. if not
supplied set the url as the env var \code{CROMWELLURL}}

\item{token}{(character) we do not recommend passing your token
here as a string. Either pass it using \code{\link[=Sys.getenv]{Sys.getenv()}} or save your
token as the env var \code{PROOF_TOKEN} and then passing nothing to this
param and we'll find it}
}
\value{
(character) the Cromwell version
}
\description{
Get the version of a Cromwell server
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwell_settings}
}

\author{
Amy Paguirigan, Scott Chamberlain
}
