% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellBackends.R
\name{cromwellBackends}
\alias{cromwellBackends}
\title{Get the available backends of a Cromwell server}
\usage{
cromwellBackends()
}
\value{
Cromwell backend options
}
\description{
Get the available backends of a Cromwell server
}
\details{
Requires valid Cromwell URL to be set in the environment.
}
\author{
Amy Paguirigan
}
