% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellCall.R
\name{cromwellCall}
\alias{cromwellCall}
\title{Retrieve metadata for the calls made in a Cromwell workflow job}
\usage{
cromwellCall(workflow_id)
}
\arguments{
\item{workflow_id}{The workflow ID to return call metadata on.}
}
\value{
Returns a long form data frame of metadata on calls. NOTE: does not currently support subWorkflows.
}
\description{
Retrieve and process call metadata for a specific workflow.
}
\details{
Requires valid Cromwell server URL to be set in the environment. (use \code{setCromwellURL()})
}
\examples{
## Request what jobs have been submitted to your Cromwell instance in the past 7 days.
recentJobs <- cromwellJobs(days = 7)
## Request workflow metadata for a specific job that was run in your Cromwell instance.
thisWorkflowID <- recentJobs$workflow_id[1]
callsMeta <- cromwellCall(workflow_id = thisWorkflowID)
}
\author{
Amy Paguirigan
}
