% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellCreate.R
\name{cromwellCreate}
\alias{cromwellCreate}
\title{Start a Cromwell server on Gizmo}
\usage{
cromwellCreate(FredHutchId = NULL, port = "2020",
  pathToServerLogs = NULL, pathToServerScript = NULL,
  pathToParams = NULL, cluster = "gizmo", local = TRUE)
}
\arguments{
\item{FredHutchId}{Your Fred Hutch username}

\item{port}{The port want to connect to for this server.}

\item{pathToServerLogs}{Full path in our file system to where you want your Cromwell server logs to be written.}

\item{pathToServerScript}{Full path in our file system to where you have saved the script used to start your Cromwell server (e.g. cromServer.sh, https://github.com/FredHutch/diy-cromwell-server).}

\item{pathToParams}{Full path in our file system to where you have saved the parameters you'd like your Cromwell server to use (e.g. cromwellParams.sh, https://github.com/FredHutch/diy-cromwell-server).}

\item{cluster}{Optional: if present, specify the location of the server, either 'gizmo' or 'beagle'}

\item{local}{Are you running this on your local machine (TRUE) or on the rhino's (FALSE)}
}
\value{
Sets your environment variable CROMWELLURL to be that of the Cromwell server you just started and returns the information about the job ID and node you'll need.
}
\description{
Start a Cromwell server on Gizmo
}
\details{
Will require your Fred Hutch Id and will prompt you to enter your Fred Hutch password.
Suggestions for parameters are:
pathToServerLogs = "/home/username/cromwell/cromwell-serverlogs/%A.txt"
pathToScript = "/home/username/cromwell/cromServer.sh"
pathToParams = "/home/username/cromwell/cromwellParams.sh"
port = "2020"
}
\author{
Amy Paguirigan
}
