% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellGlob.R
\name{cromwellGlob}
\alias{cromwellGlob}
\title{Pull a glob of metadata for a specific Cromwell workflow job}
\usage{
cromwellGlob(workflow_id, expandSubWorkflows = F)
}
\arguments{
\item{workflow_id}{The workflow ID to return metadata for.}

\item{expandSubWorkflows}{Boolean, whether to expand subworkflows in the results or not, default is FALSE.}
}
\value{
Returns a gross list of lists of metadata on a workflow.
}
\description{
Retrieve a glob of workflow level metadata for a specific workflow.
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwellSettings}
}

\examples{
\dontrun{
## Request what jobs have been submitted to your Cromwell instance in the past 7 days.
recentJobs <- cromwellJobs(days = 7)
## Request workflow metadata for a specific job that was run in your Cromwell instance.
thisWorkflowID <- recentJobs$workflow_id[1]
workflowMeta <- cromwellGlob(workflow_id = thisWorkflowID)
}
}
\author{
Amy Paguirigan
}
