% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellOutputs.R
\name{cromwellOutputs}
\alias{cromwellOutputs}
\title{Gets outputs for a workflow in Cromwell}
\usage{
cromwellOutputs(
  workflow_id,
  cromURL = Sys.getenv("CROMWELLURL", unset = "needsURL")
)
}
\arguments{
\item{workflow_id}{Unique workflow id of the job.}

\item{cromURL}{The full string of the Cromwell URL to query if not using this locally (e.g. http://gizmog10:8000). (Optional)}
}
\value{
Returns a dataframe containing the workflowName, workflowOutputType,
pathToOutput, shardIndex for the specified outputs of a workflow_id
}
\description{
Gets outputs for a workflow in Cromwell
}
\details{
Requires valid Cromwell server URL to be set in the environment, or the use
of the cromURL param if you want to specify upon call the URL to use. (use \code{setCromwellURL()})
}
\author{
Amy Paguirigan
}
