% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellTiming.R
\name{cromwellTiming}
\alias{cromwellTiming}
\title{Get a timing diagram for a Cromwell workflow}
\usage{
cromwellTiming(
  workflow_id,
  cromURL = Sys.getenv("CROMWELLURL", unset = "needsURL")
)
}
\arguments{
\item{workflow_id}{Unique workflow id of the job.}

\item{cromURL}{The full string of the Cromwell URL to query if not using this locally (e.g. http://gizmog10:8000). (Optional)}
}
\value{
Opens a timing diagram in a browser
}
\description{
Get a timing diagram for a Cromwell workflow
}
\details{
Requires valid Cromwell server URL to be set in the environment, or the use
of the cromURL param if you want to specify upon call the URL to use. (use \code{setCromwellURL()})
}
\examples{
TBD
}
\author{
Amy Paguirigan
}
