% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellVersion.R
\name{cromwellVersion}
\alias{cromwellVersion}
\title{Get the version of a Cromwell server}
\usage{
cromwellVersion()
}
\value{
(character) the Cromwell version
}
\description{
Get the version of a Cromwell server
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwellSettings}
}

\author{
Amy Paguirigan
}
