% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{cw_url}
\alias{cw_url}
\title{Look for Cromwell URL as the env var CROMWELLURL}
\usage{
cw_url()
}
\value{
a string, empty if env var not found
}
\description{
Look for Cromwell URL as the env var CROMWELLURL
}
