% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAWSLogs.R
\name{getAWSLogs}
\alias{getAWSLogs}
\title{Get AWS job logs}
\usage{
getAWSLogs(jobId = NULL)
}
\arguments{
\item{jobId}{JobId from AWS of a job to request logs for from a Crowmell workflow}
}
\value{
A list of two entries: one the AWS Cloudwatch Logs (log), and one the AWS Batch job metadata (Job) for the jobId submitted.
}
\description{
Get AWS job logs
}
\details{
Requires valid Cromwell URL to be set in the environment and valid AWS credentials as well.
}
\examples{
TBD
}
\author{
Amy Paguirigan
}
