% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCromwellURL.R
\name{setCromwellURL}
\alias{setCromwellURL}
\title{Set the URL for your Cromwell server's API}
\usage{
setCromwellURL(nodeAndPort = NULL)
}
\arguments{
\item{nodeAndPort}{Your node and port of the running Cromwell server on Gizmo (e.g., gizmoz85:12345)}
}
\value{
Sets your environment variable CROMWELLURL to be that of the Cromwell server for the job specified.
}
\description{
Set the URL for your Cromwell server's API
}
\author{
Amy Paguirigan
}
