% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{try_auth_header}
\alias{try_auth_header}
\title{Add Authorization header}
\usage{
try_auth_header(token = NULL)
}
\arguments{
\item{token}{(character) a Bearer token. optional. if nothing
passed, we look for the env var PROOF_TOKEN}
}
\value{
A \code{request} S3 class with the HTTP header that can be passed
to \code{httr::GET()}, \code{httr::POST()}, etc.
}
\description{
Add Authorization header
}
\keyword{internal}
