% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{try_proof_header}
\alias{try_proof_header}
\title{Get header for PROOF API calls}
\usage{
try_proof_header()
}
\value{
A \code{request} S3 class with the HTTP header that can be passed
to \code{httr::GET()}, \code{httr::POST()}, etc.
}
\description{
Get header for PROOF API calls
}
\keyword{internal}
