% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromUtils.R
\name{workflow_options}
\alias{workflow_options}
\title{Pull the workflow options provided for a Cromwell workflow job}
\usage{
workflow_options(workflow_id, url = cw_url(), token = NULL)
}
\arguments{
\item{workflow_id}{(character) A single workflow ID}

\item{url}{(character) base url for your Cromwell server. optional. if not
supplied set the url as the env var \code{CROMWELLURL}}

\item{token}{(character) we do not recommend passing your token
here as a string. Either pass it using \code{\link[=Sys.getenv]{Sys.getenv()}} or save your
token as the env var \code{PROOF_TOKEN} and then passing nothing to this
param and we'll find it}
}
\value{
Returns a data frame of the options for a workflow previously run
}
\description{
Pull the workflow options provided for a Cromwell workflow job
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwell_settings}
}

\examples{
\dontrun{
jobs <- cromwell_jobs()
workflow_options(jobs$workflow_id[1])
}
}
\author{
Amy Paguirigan, Scott Chamberlain
}
