% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromUtils.R
\name{workflowOptions}
\alias{workflowOptions}
\title{Pull the workflow options provided for a Cromwell workflow job}
\usage{
workflowOptions(
  workflow_id,
  cromURL = Sys.getenv("CROMWELLURL", unset = "needsURL")
)
}
\arguments{
\item{workflow_id}{The workflow ID to return options for.}

\item{cromURL}{The full string of the Cromwell URL to query if not using this locally (e.g. http://gizmog10:8000). (Optional)}
}
\value{
Returns a data frame of the options for a workflow previously run
}
\description{
Pull the workflow options provided for a Cromwell workflow job
}
\details{
Requires valid Cromwell server URL to be set in the environment, or the use
of the cromURL param if you want to specify upon call the URL to use. (use \code{setCromwellURL()})
}
\author{
Amy Paguirigan
}
