% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellAbort.R
\name{cromwell_abort}
\alias{cromwell_abort}
\title{Abort a workflow job on Cromwell}
\usage{
cromwell_abort(workflow_id, url = cw_url(), token = NULL)
}
\arguments{
\item{workflow_id}{(character) A single workflow ID}

\item{url}{(character) base url for your Cromwell server. optional. if not
supplied set the url as the env var \code{CROMWELLURL}}

\item{token}{(character) we do not recommend passing your token
here as a string. Either pass it using \code{\link[=Sys.getenv]{Sys.getenv()}} or save your
token as the env var \code{PROOF_TOKEN} and then passing nothing to this
param and we'll find it}
}
\value{
a tibble
}
\description{
Abort a workflow job on Cromwell
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwell_settings}
}

\author{
Amy Paguirigan, Scott Chamberlain
}
