% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellBackends.R
\name{cromwell_backends}
\alias{cromwell_backends}
\title{Get the available backends of a Cromwell server}
\usage{
cromwell_backends()
}
\value{
Cromwell backend options
}
\description{
Get the available backends of a Cromwell server
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwell_settings}
}

\author{
Amy Paguirigan, Scott Chamberlain
}
