% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellCache.R
\name{cromwell_cache}
\alias{cromwell_cache}
\title{Pull Cromwell Call Caching Data}
\usage{
cromwell_cache(workflow_id)
}
\arguments{
\item{workflow_id}{(character) A single workflow ID}
}
\value{
Returns a long form data frame of metadata on call caching in a
workflow. NOTE: Currently does not support subworkflows well.
}
\description{
Gets info about call caching status for the calls of a workflow
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwell_settings}
}

\author{
Amy Paguirigan, Scott Chamberlain
}
