% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellCache.R
\name{cromwell_cache}
\alias{cromwell_cache}
\title{Pull Cromwell Call Caching Data}
\usage{
cromwell_cache(workflow_id, url = cw_url(), token = NULL)
}
\arguments{
\item{workflow_id}{(character) A single workflow ID}

\item{url}{(character) base url for your Cromwell server. optional. if not
supplied set the url as the env var \code{CROMWELLURL}}

\item{token}{(character) we do not recommend passing your token
here as a string. Either pass it using \code{\link[=Sys.getenv]{Sys.getenv()}} or save your
token as the env var \code{PROOF_TOKEN} and then passing nothing to this
param and we'll find it}
}
\value{
Returns a long form data frame of metadata on call caching in a
workflow. NOTE: Currently does not support subworkflows well.
}
\description{
Gets info about call caching status for the calls of a workflow
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwell_settings}
}

\author{
Amy Paguirigan, Scott Chamberlain
}
