% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellGlob.R
\name{cromwell_glob}
\alias{cromwell_glob}
\title{Pull a glob of metadata for a specific Cromwell workflow job}
\usage{
cromwell_glob(workflow_id, expand_sub_workflows = FALSE)
}
\arguments{
\item{workflow_id}{(character) A single workflow ID}

\item{expand_sub_workflows}{Boolean, whether to expand subworkflows in the
results or not, default is FALSE.}
}
\value{
Returns a gross list of lists of metadata on a workflow.
}
\description{
Retrieve a glob of workflow level metadata for a specific workflow.
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwellSettings}
}

\examples{
\dontrun{
## Request what jobs have been submitted to your Cromwell instance in the
## past 7 days.
recentJobs <- cromwell_jobs(days = 7)
## Request workflow metadata for a specific job that was run in your
## Cromwell instance.
thisWorkflowID <- recentJobs$workflow_id[1]
workflowMeta <- cromwell_glob(workflow_id = thisWorkflowID)
}
}
\author{
Amy Paguirigan
}
