% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellJobs.R
\name{cromwell_jobs}
\alias{cromwell_jobs}
\title{Requests metadata about Cromwell workflow jobs during a time period specified}
\usage{
cromwell_jobs(days = 1, workflow_name = NULL, workflow_status = NULL)
}
\arguments{
\item{days}{The number of days of history to return, defaults to 1 day.}

\item{workflow_name}{An array of strings of valid workflow names you want in
your job list.}

\item{workflow_status}{A array of strings of valid workflow statuses you want
in your job list (e.g., submitted, running, succeeded, failed,
aborting, aborted)}
}
\value{
Returns a long form data frame of metadata on workflow jobs
submitted to a specific Cromwell instance.
}
\description{
Requests metadata about Cromwell workflow jobs during a time period specified
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwellSettings}
}

\examples{
\dontrun{
## Request what jobs have been submitted to your Cromwell instance in the
## past 7 days.
recentJobs <- cromwell_jobs(days = 7)
}
}
\author{
Amy Paguirigan
}
