% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellLogs.R
\name{cromwell_logs}
\alias{cromwell_logs}
\title{Gets logs for a workflow in Cromwell}
\usage{
cromwell_logs(workflow_id)
}
\arguments{
\item{workflow_id}{(character) A single workflow ID}
}
\value{
Returns the response from the API post
}
\description{
Gets logs for a workflow in Cromwell
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwell_settings}
}

\author{
Amy Paguirigan, Scott Chamberlain
}
