% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellSubmitBatch.R
\name{cromwell_submit_batch}
\alias{cromwell_submit_batch}
\title{Submit a workflow job to Cromwell}
\usage{
cromwell_submit_batch(
  wdl,
  batch = NULL,
  params = NULL,
  options = NULL,
  labels = NULL,
  dependencies = NULL
)
}
\arguments{
\item{wdl}{Local path to the wdl file describing the workflow. (Required)}

\item{batch}{Local path to the json containing a reference to any batch file
desired if the workflow is a batch. (Optional)}

\item{params}{Local path to the json containing the parameters to use with
the workflow. (Optional)}

\item{options}{Local path to the json containing workflow options to
apply. (Optional)}

\item{labels}{A data frame containing the labels for this workflow.
(Optional)}

\item{dependencies}{A zip'd file of subworkflow dependencies. (Optional)}
}
\value{
Returns the response from the API post which includes the workflow
ID that you'll need to monitor the job.
}
\description{
Supports the submission of a fully defined workflow job to a Cromwell
instance.
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwell_settings}
}

\author{
Amy Paguirigan, Scott Chamberlain
}
