% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellTiming.R
\name{cromwell_timing}
\alias{cromwell_timing}
\title{Get a timing diagram for a Cromwell workflow}
\usage{
cromwell_timing(workflow_id, url = cw_url())
}
\arguments{
\item{workflow_id}{(character) A single workflow ID}

\item{url}{(character) base url for your Cromwell server. optional. if not
supplied set the url as the env var \code{CROMWELLURL}}
}
\value{
Opens a timing diagram in a browser
}
\description{
Get a timing diagram for a Cromwell workflow
}
\details{
Internally this function uses \code{\link[httr:BROWSE]{httr::BROWSE()}} which uses
\code{\link[utils:browseURL]{utils::browseURL()}} - if a auth header is required you'll have to
do that manually
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwell_settings}
}

\author{
Amy Paguirigan, Scott Chamberlain
}
