% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellTiming.R
\name{cromwell_timing}
\alias{cromwell_timing}
\title{Get a timing diagram for a Cromwell workflow}
\usage{
cromwell_timing(workflow_id)
}
\arguments{
\item{workflow_id}{(character) A single workflow ID}
}
\value{
Opens a timing diagram in a browser
}
\description{
Get a timing diagram for a Cromwell workflow
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwellSettings}
}

\author{
Amy Paguirigan
}
