% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellVersion.R
\name{cromwell_version}
\alias{cromwell_version}
\title{Get the version of a Cromwell server}
\usage{
cromwell_version()
}
\value{
(character) the Cromwell version
}
\description{
Get the version of a Cromwell server
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwellSettings}
}

\author{
Amy Paguirigan
}
