% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellAbort.R
\name{cromwellAbort}
\alias{cromwellAbort}
\title{Abort a workflow job on Cromwell}
\usage{
cromwellAbort(workflow_id)
}
\arguments{
\item{workflow_id}{Unique workflow id of the job you wish to kill.}
}
\value{
Returns the response from the API post
}
\description{
Aborts any given workflow job on Cromwell.
}
\details{
Requires valid Cromwell URL to be set in the environment.
}
\author{
Amy Paguirigan
}
