% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellBackends.R
\name{cromwellBackends}
\alias{cromwellBackends}
\title{Get the available backends of a Cromwell server}
\usage{
cromwellBackends(cromURL = Sys.getenv("CROMWELLURL", unset = "needsURL"))
}
\arguments{
\item{cromURL}{The full string of the Cromwell URL to query (e.g. http://gizmog10:8000). (Optional)}
}
\value{
Cromwell backend options
}
\description{
Get the available backends of a Cromwell server
}
\details{
Requires valid Cromwell server URL to be set in the environment, or the use
of the cromURL param if you want to specify upon call the URL to use.
}
\author{
Amy Paguirigan
}
