% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellCache.R
\name{cromwellCache}
\alias{cromwellCache}
\title{Pull Cromwell Call Caching Data}
\usage{
cromwellCache(workflow_id)
}
\arguments{
\item{workflow_id}{The workflow ID to return call caching metadata for.}
}
\value{
Returns a long form data frame of metadata on call caching in a workflow. NOTE: Currently does not support subworkflows.
}
\description{
Gets info about call caching status for the calls of a workflow
}
\details{
Requires valid Cromwell server URL to be set in the environment. (use \code{setCromwellURL()})
}
\author{
Amy Paguirigan
}
