% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellJobs.R
\name{cromwellJobs}
\alias{cromwellJobs}
\title{Pull information about recent Cromwell workflow jobs}
\usage{
cromwellJobs(days = 1)
}
\arguments{
\item{days}{The number of days of history to return, defaults to 1 day.}
}
\value{
Returns a long form data frame of metadata on workflow jobs submitted to a specific Cromwell instance.
}
\description{
Requests metadata about Cromwell workflow jobs during a time period specified.
}
\details{
Requires valid Cromwell server URL to be set in the environment. (use \code{setCromwellURL()})
}
\examples{
## Request what jobs have been submitted to your Cromwell instance in the past 7 days.
recentJobs <- cromwellJobs(days = 7)
}
\author{
Amy Paguirigan
}
