% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellOutputs.R
\name{cromwellOutputs}
\alias{cromwellOutputs}
\title{Gets outputs for a workflow in Cromwell}
\usage{
cromwellOutputs(workflow_id)
}
\arguments{
\item{workflow_id}{Unique workflow id of the job.}
}
\value{
Returns the response from the API post
}
\description{
Gets outputs for a workflow in Cromwell
}
\details{
Requires valid Cromwell URL to be set in the environment.
}
\author{
Amy Paguirigan
}
