% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellOutputs.R
\name{cromwellOutputs}
\alias{cromwellOutputs}
\title{Gets outputs for a workflow in Cromwell}
\usage{
cromwellOutputs(workflow_id)
}
\arguments{
\item{workflow_id}{(character) A single workflow ID}
}
\value{
Returns a dataframe containing the workflowName, workflowOutputType,
pathToOutput, shardIndex for the specified outputs of a workflow_id
}
\description{
Gets outputs for a workflow in Cromwell
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwellSettings}
}

\author{
Amy Paguirigan
}
