% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellSubmitBatch.R
\name{cromwellSubmitBatch}
\alias{cromwellSubmitBatch}
\title{Submit a workflow job to Cromwell}
\usage{
cromwellSubmitBatch(WDL, Batch, Params = NULL, Options = NULL,
  Labels = NULL, Dependencies = NULL)
}
\arguments{
\item{WDL}{Local path to the wdl file describing the workflow. (Required)}

\item{Batch}{Local path to the json containing a reference to any batch file desired if the workflow is a batch. (Optional)}

\item{Params}{Local path to the json containing the parameters to use with the workflow. (Optional)}

\item{Options}{Local path to the json containing workflow options to apply.(Optional)}

\item{Labels}{A data frame containing the labels for this workflow.(Optional)}

\item{Dependencies}{A zip'd file of subworkflow dependencies. (Optional)}
}
\value{
Returns the response from the API post which includes the workflow ID that you'll need to monitor the job.
}
\description{
Supports the submission of a fully defined workflow job to a Cromwell instance.
}
\details{
Requires valid Cromwell server URL to be set in the environment. (use \code{setCromwellURL()})
}
\examples{
TBD
}
\author{
Amy Paguirigan
}
