% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellTiming.R
\name{cromwellTiming}
\alias{cromwellTiming}
\title{Get a timing diagram for a Cromwell workflow}
\usage{
cromwellTiming(workflow_id)
}
\arguments{
\item{workflow_id}{Unique workflow id of the job.}
}
\value{
Opens a timing diagram in a browser
}
\description{
Get a timing diagram for a Cromwell workflow
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwellSettings}
}

\author{
Amy Paguirigan
}
