% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellValidate.R
\name{cromwellValidate}
\alias{cromwellValidate}
\title{Validates a workflow for submission}
\usage{
cromwellValidate(wdl, all_inputs = NULL)
}
\arguments{
\item{wdl}{Local path to the wdl file describing the workflow. (Required)}

\item{all_inputs}{Local path to the json containing ALL the inputs the
parameters to use with the workflow - if a batch and parameters json
are present they must first be combined. (Optional)}
}
\value{
Returns the response from the API post which includes the workflow
ID that you'll need to monitor the job.
}
\description{
Validates a workflow for submission
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwellSettings}
}

\author{
Amy Paguirigan
}
