% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellVersion.R
\name{cromwellVersion}
\alias{cromwellVersion}
\title{Get the version of a Cromwell server}
\usage{
cromwellVersion()
}
\value{
Cromwell version
}
\description{
Get the version of a Cromwell server
}
\details{
Requires valid Cromwell URL to be set in the environment.
}
\examples{
TBD
}
\author{
Amy Paguirigan
}
