% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellVersion.R
\name{cromwellVersion}
\alias{cromwellVersion}
\title{Get the version of a Cromwell server}
\usage{
cromwellVersion(cromURL = Sys.getenv("CROMWELLURL", unset = "needsURL"))
}
\arguments{
\item{cromURL}{The full string of the Cromwell URL to query  (e.g. http://gizmog10:8000). (Optional)}
}
\value{
Cromwell version
}
\description{
Get the version of a Cromwell server
}
\details{
Requires valid Cromwell server URL to be set in the environment or specify "cromURL".
}
\author{
Amy Paguirigan
}
