% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromwellWorkflow.R
\name{cromwellWorkflow}
\alias{cromwellWorkflow}
\title{Pull metadata for a specific Cromwell workflow job}
\usage{
cromwellWorkflow(
  workflow_id,
  cromURL = Sys.getenv("CROMWELLURL", unset = "needsURL")
)
}
\arguments{
\item{workflow_id}{The workflow ID to return metadata for.}

\item{cromURL}{The full string of the Cromwell URL to query (e.g. http://gizmog10:8000). (Optional)}
}
\value{
Returns a long form data frame of metadata on a workflow.
}
\description{
Retrieve and process all labels, submission and workflow level metadata for a specific workflow.
}
\details{
Requires valid Cromwell server URL to be set in the environment, or the use
of the cromURL param if you want to specify upon call the URL to use.
}
\examples{
\dontrun{
## Request what jobs have been submitted to your Cromwell instance in the past 7 days.
recentJobs <- cromwellJobs(days = 7)
## Request workflow metadata for a specific job that was run in your Cromwell instance.
thisWorkflowID <- recentJobs$workflow_id[1]
workflowMeta <- cromwellWorkflow(workflow_id = thisWorkflowID)
}
}
\author{
Amy Paguirigan
}
