% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcromwell-package.R
\docType{package}
\name{rcromwell-package}
\alias{rcromwell}
\alias{rcromwell-package}
\title{rcromwell: Convenience Tools for Managing WDL Workflows via Cromwell}
\description{
A repo containing a basic R package for using Cromwell with WDL workflows via R.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://getwilds.org/rcromwell/ (website) https://github.com/getwilds/rcromwell (devel)}
  \item Report bugs at \url{https://github.com/getwilds/rcromwell/issues}
}

}
\author{
\strong{Maintainer}: Amy Paguirigan \email{apaguiri@fredhutch.org} (\href{https://orcid.org/0000-0002-6819-9736}{ORCID})

}
\keyword{internal}
