% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcromwell-package.R
\docType{package}
\name{rcromwell-package}
\alias{rcromwell}
\alias{rcromwell-package}
\title{rcromwell: Convenience Tools for Managing WDL Workflows via Cromwell}
\description{
A repo containing a basic R package for using Cromwell with WDL workflows via R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://getwilds.org/rcromwell}
  \item \url{https://github.com/getwilds/rcromwell}
  \item Report bugs at \url{https://github.com/getwilds/rcromwell/issues}
}

}
\author{
\strong{Maintainer}: Scott Chamberlain \email{sachamber@fredhutch.org} (\href{https://orcid.org/0000-0003-1444-9135}{ORCID})

Authors:
\itemize{
  \item Amy Paguirigan (\href{https://orcid.org/0000-0002-6819-9736}{ORCID})
}

Other contributors:
\itemize{
  \item Fred Hutchinson Cancer Center \email{wilds@fredhutch.org} [funder]
}

}
\keyword{internal}
