% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{try_auth_header}
\alias{try_auth_header}
\title{Add Authorization header}
\usage{
try_auth_header(.req, token = NULL)
}
\arguments{
\item{.req}{An \code{httr2} request}

\item{token}{(character) a Bearer token. optional. if nothing
passed, we look for the env var PROOF_TOKEN}
}
\value{
An \code{httr2_request} S3 class adding an HTTP header for
\code{Authorization} with the value in \code{token}
}
\description{
Add Authorization header
}
\keyword{internal}
