% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromUtils.R
\name{workflow_options}
\alias{workflow_options}
\title{Pull the workflow options provided for a Cromwell workflow job}
\usage{
workflow_options(workflow_id)
}
\arguments{
\item{workflow_id}{(character) A single workflow ID}
}
\value{
Returns a data frame of the options for a workflow previously run
}
\description{
Pull the workflow options provided for a Cromwell workflow job
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwell_settings}
}

\examples{
\dontrun{
jobs <- cromwell_jobs()
workflow_options(jobs$workflow_id[1])
}
}
\author{
Amy Paguirigan, Scott Chamberlain
}
