% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromUtils.R
\name{workflowInputs}
\alias{workflowInputs}
\title{Pull the workflow inputs provided for a Cromwell workflow job}
\usage{
workflowInputs(workflow_id)
}
\arguments{
\item{workflow_id}{The workflow ID to return inputs for.}
}
\value{
Returns a data frame of the inputs for a workflow previously run
}
\description{
Pull the workflow inputs provided for a Cromwell workflow job
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwellSettings}
}

\author{
Amy Paguirigan
}
