% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cromUtils.R
\name{workflowOptions}
\alias{workflowOptions}
\title{Pull the workflow options provided for a Cromwell workflow job}
\usage{
workflowOptions(workflow_id)
}
\arguments{
\item{workflow_id}{The workflow ID to return options for.}
}
\value{
Returns a data frame of the options for a workflow previously run
}
\description{
Pull the workflow options provided for a Cromwell workflow job
}
\section{Important}{

Requires valid Cromwell server URL to be set. See \link{cromwellSettings}
}

\author{
Amy Paguirigan
}
